Closes #1852: Add small string groupby perf to `run_benchmarks` (#1853)
* Closes #1852: Add small string groupby perf to `run_benchmarks`

This PR (closes #1852) adds small string groupby perf to `run_benchmarks` so it will get tested in the nightly.

The PR feedback on #1848 mentioned `arkouda-string.graph` might be a more appropriate place for this benchmark rathen than it's own graph. I considered having it there originally, but I felt like it wasn't really comparable to the other keys there since it's comparing how groupby scales with multiple arrays 1,2,8,16 and this is tracking the performance based on the size of the strings within the SegString. That being said, I don't have terribly strong opinions on this and would be happy to change it if others feel that's the way to go

* move to `arkouda-string.graph`

Co-authored-by: Pierce Hayes <pierce314159@users.noreply.github.com>