Swap order of zippered iterations in which an unbounded range was the leader
In Chapel, we recently fixed an inconsistency between serial and
parallel loops such that 'for (i,j) in zip(1.., whatever)' now results
in a (conceptually) infinite execution because we make the first
iterand determine the number of iterations of the loop, and subsequent
iterands "follow" the first's lead.  Prior to this change, if the
first iterand was an unbounded loop, it would just stop whenever the
other iterators were complete.  This PR flips such cases around in the
source to 'for (j,i) in zip(whatever, 1..)' to preserve the old
behavior in the face of the new interpretation.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>