Closes 1943: Quiet deprecation warnings in prep for Chapel 1.29 (#1944)
* Avoid using `channel` type

`channel` has been deprecated in favor of `fileReader`/`fileWriter`, but
we don't actually need to refer to the type so just avoid it.

* Avoid using `Array.sorted`

It's been deprecated in favor of just calling `sort(Array)`

* Avoid using `Timer`

It's been deprecated in favor of `stopwatch`, but most places in arkouda
use `getCurrentTime()` so I just switched to that style for the few
places using a timer.

* Avoid using `opentmp`

It's been deprecated in favor of `openTempFile`, but we didn't actually
need a file here, so just use `openmem` instead.

* Avoid using `createVersion()` since it's deprecated

`createVersion()` is deprecated in favor of `new versionValue()`. This
version check was being used as a compatibility shim, so just create a
real compat shim for appending to the end of an existing file instead of
doing a version check in source.