Add aggregation interface for `bigint`
This does **not** actually aggregate `bigint`, it just adds the
interface so we can start writing code to use it. This is a code clone
of the normal aggregators with plain assignments added in. I want the
code clone so it's easier to parse the diff when actual aggregation
support is added. This also turns unorderedCopy into plain assignment
for non-POD types (like bigint) since unorderedCopy only works for POD.

Part of 1994