Add DstAggregator support for `bigint`
Add a DstAggregator that supports bigint. Plain old integer aggregators
locally aggregate the target address and the inline value by storing a
tuple of `(c_ptr(int), int)` but it's not as simple with variable sized
bigints. It's also not as easy to extract the raw value since bigints
have metadata stored inline, but the values stored off in the heap. To
aggregate this our buffers are no longer typed and are just raw bytes
that we'll interpret. We now locally aggregate the target address, the
8-byte sign+size (int64 that stores the number of limbs and the sign of
the limbs), and the variables sized limbs. This uses internal functions
to expose the internal sign+size and limbs. With this we have our target
address and "value" stored inline and can get data to the target side
like normal. On the target side we then extract the address and size,
and then resize the target bigint, copy the limbs over, and update the
target sign+size.

Implementation TODOs:
 - Fix support for bigints with more limbs than we can store
 - Fix support for mismatched bigint record and mpz locale.id
 - Tune buffer size (currently sized to fit the same number of elements
   as plain old int64 aggregation to do comm comparisons)
 - See if we can use public functions (`mpz_limbs_write`,
   `mpz_limbs_finish`) added in GMP 6 --
   https://gmplib.org/manual/Integer-Special-Functions

Testing TODOs:
 - check negative numbers (gmp stores sign as part of size)
 - check massive numbers (more limbs than we have buffer space for,
   broken currently)
 - check variable sized limbs (only tested with single limb so far)
 - check performance with larger limbs (single limb is ~2x slower)
 - check with bigint record wrapper and heap-allocated mpz on different
   nodes (broken currently)
 - check with limb sizes that will grow and shrink target bigint
 - check with aggregating 0 values (no limbs). Should work with current
   strategy, but would require special case with `mpz_limbs_write`