Use Chapel runtime view of allocatable memory when available
Historically, Arkouda has just used the amount of physical memory to
determine when an operation may exceed memory capacity, but this isn't
always and accurate representation of how much memory can be allocated.
When a fixed heap is being used, ask the runtime how large the heap is
since that represents the max amount of memory that can be allocated.

Currently, this uses an internal runtime call but I'll look to see if
something like this could be added to the standard library.

Part of 2044