Create task private copies of values for pdarray=value (#2208)
Now that bigint is using the shared code with the other types for
indexing implemented in #2081, there was a bug uncovered in the
bigint module failing with remote assignments that will be fixed
in 1.30. For 1.29, using task-private copies is a suitable
workaround to avoid the remote assignmnets.

Also, this will result in a performance improvement for bigint
indexing, as discussed in
https://github.com/Bears-R-Us/arkouda/pull/2157#discussion_r1110440741.