Add initial aggregation support for `bigint` (#2290)
Plain old integer DstAggregators locally aggregate the target address
and the inline value by storing a tuple of `(c_ptr(int), int)` but it's
not as simple with variable sized bigints. It's also not as easy to
extract the raw value since bigints have metadata stored inline, but the
values stored off in the heap. To enable aggregation, the buffers are no
longer typed and are just raw bytes that we'll interpret. We now locally
aggregate the target address, the 8-byte sign+size (int64 that stores
the number of limbs and the sign of the limbs), and the variable sized
limbs. This uses internal functions to expose the internal sign+size and
limbs. With this we have our target address and "value" stored inline
and can get data to the target side like normal. On the target side we
then extract the address and size, and then resize the target bigint,
copy the limbs over, and update the target sign+size.

SrcAggregator is a little more complicated since we don't know how big
our src values are until we're on the remote node. Because of this we
can't pre-allocate space for all our elements, so instead we just
aggregate addresses locally and when those buffers are full we fill up
value buffers as much as we can and just loop until we've processed all
values.

This significantly improves the performance of operations that were
using the aggregation interface. Testing with single limb (64-bit value)
bigints, I see aggregation performance that's within 2x of plain old
int(64) aggregation. This is around what I'd expect since we have to
aggregate an extra 8 bytes for the sign_size and there's other overheads
with real location and extracting limbs. Here's perforamnce for bigint
groupby (which does a bigint gather) on 16-node-cs-hdr:

```
>>> arkouda bigint groupby
numLocales = 16, N = 16,000,000
2-array Average time = 16.3907 sec
2-array Average rate = 0.0145 GiB/sec

>>> arkouda bigint groupby
numLocales = 16, N = 16,000,000
2-array Average time = 0.9008 sec
2-array Average rate = 0.2647 GiB/sec

>>> arkouda bigint groupby
numLocales = 16, N = 1,600,000,000
2-array Average time = 4.0616 sec
2-array Average rate = 5.8700 GiB/sec
```

I don't think there's much more performance we can squeeze out of this,
but there is certainly some code cleanup and more testing I'd like to
add in the future. That's captured in 2288

Resolves 1994
Resolves 2212