Eliminate some communication from datetime methods
Previously, helper arrays like `MONTHOFFSET` required communication
from non-0 locales because they were arrays declared on locale 0. This
resulted in a significant amount of communication for datetime methods
running on non-0 locales.

Here, switch helper arrays to tuples, which eliminates comm because the
Chapel compiler knows how to serialize tuples and since they're consts
that allows Chapel to broadcast/replicate the tuples to remote locales a
program startup and later uses will refer to the replicated locale
private version instead of doing communication. This is similar to what
was done upstream in chapel-lang/chapel 22050 for helper arrays in the
datetime module. Ideally, Chapel would be able to replicate/broadcast
small const arrays too, but it doesn't have support for that today.

Part of 2317