Closes #2304: `inner_join` on `Strings` and `Categorical` (#2453)
* Closes #2304: Broadcast on Strings and Categorical

This PR (closes #2304) adds broadcasting for Strings and inner join on Strings and Categorical. We can't add broadcasting to Categorical since we use `broadcast` in the init of Categorical so this causes circular dependecies. Otherwise we could just broadcast on the codes and build `from_codes`

This is built upon the work in #2376

Co-authored-by: joshmarshall1 <josh@MSI.localdomain>
Co-authored-by: Pierce Hayes <pierce314159@users.noreply.github.com>

* Adds logic for perm broadcast and adds tests for both versions of broadcast

* Update src/BroadcastMsg.chpl

---------

Co-authored-by: joshmarshall1 <josh@MSI.localdomain>
Co-authored-by: Pierce Hayes <pierce314159@users.noreply.github.com>