Clang 15 changed the default PIE setting on some systems (#2468)
(https://releases.llvm.org/15.0.0/tools/clang/docs/ReleaseNotes.html),
which can cause a problem when compiling the Arrow executable. In the
Makefile today, `CXX` is used to compile the Arrow code, but that
does not always line up with Chapel's C++ compiler, which could
mean that the Arrow code may be using an older version of clang,
which could throw the PIE versions out of sync.

This PR gets the Chapel C++ compiler in the Makefile and uses that
to compile Arrow so that they won't have mismatched compilers.