Adjust getUserMetrics to return borrowed
This PR is intended to fix Arkouda compilation failures after the Chapel
PR https://github.com/chapel-lang/chapel/pull/21894 .

In particular, there is a change in behavior for statements like this:

     var userMetrics : borrowed CounterTable = this.getUserMetrics(...);

because after that PR, the 'shared' result of getUserMetrics is
destroyed at the end of the statement, before 'userMetrics' can be used.

This PR changes getUserMetrics to return a 'borrowed CounterTable' which
seems to better match how it is used in other places in the module and
addresses the issue.