Fixes #2627: Failures with `flake8 6.1.0`
This PR fixes #2627.

flake8 latest version, 6.1.0, was release on 7/29. This version update was casuing the CI to fail on master. This adds the changes needed for arkouda to pass the lastes version of flake8.

We should all be sure to update our conda env to have the latest version. For me the easiest way to do this was to blow away the old env and recreate from the yaml file. It might be necessarry to update your conda channels