Fix alignment in aggregation allocation (#2678)
In https://github.com/Bears-R-Us/arkouda/pull/2489, I had replaced
calls to `c_aligned_alloc` to `allocate` and accidentally swapped
the position of the padding and the number of locales in one spot,
so reverting that here.

This was seen to cause crashes at 512 locales and the fix was
identified by @ronawho.

See: https://github.com/Cray/chapel-private/issues/5247