Fix bug with naming after new registry refactor (#2722)
unregister an array when that `name` field isn't set when creating
registry entries. To get around that, in the initializers, the name
field needs to be set. This PR handles that for the `ArrayRegEntry`,
which was causing testing failures, but there are other types that
have multiple names where it isn't as clear what "name" should be.
I suspect those cases need to be fixed as well.