Fix last outstanding error regarding OOM handling compat modules (#2763)
In #2760, the compat modules for the 1.32 release switched to
using the already created domain to create an array, rather than
building the domain from scratch, which was causing some overhead.
What that missed was the case where there may be a DR array passed
in when you would like to create a block distributed array, so a
check on the type of the domain is being added here.