Fixes #2804: Missing values in CSV bug
This PR (fixes #2804) missing values are interpreted as the empty string, which causes a bad cast when the column dtype `!= string`. This PR throws and `IOError` for missing values in a CSV file unless the `allow_errors` flag is set

If `allow_errors` is True, then any missing values are replaced with `min(col_dtype)`. In the future we can look into converting the column to a float and using `nan`s to match pandas, but we don't know if there's missing values in a column until we've already started reading. So I don't think there's gonna be an efficient way to do this

Working this bug uncovered an issue when an error is thrown in a parallel loop, this is being investigated further in https://github.com/chapel-lang/chapel/issues/23648. This PR works around this by using a boolean `hadError` and an or reduction across all the tasks, so we can throw the error after the parallel loop has completed