Fixes #2816: Resolve nil config file bug (#2817)
This PR fixes #2816. When launching the server using chpl 1.32 with gasnet enabled, a file with the name `(nil)` or `0x0` was created containing the server connection info. Some change from 1.32 seems to have caused `getEnv`'s return value to change. This was resolved by doing a `nil` check along with the empty string check

Co-authored-by: Pierce Hayes <pierce314159@users.noreply.github.com>