Add support for reading Decimal128 Parquet columns (#2821)
* Add support for reading Decimal128 Parquet columns
Previously, when reading a file with a Decimal128 column, Arkouda
would silently ignore the column and the data would be dropped.
This PR adds the ability to read Decimal128 Parquet columns,
but they are cast to float C++ values and stored into a 64-bit
`real` Chapel array, so some precision can be lost in the
conversion.

Longer term, we would like to add support for arbitrary-precision
values into Chapel, which would result in a new type, but that
would be able to more accurately display Decimal128 values.

* Add lookup table for precision values to calculate byte length

* Add test

* Add comment

* Remove unnecessary batch size check based off Pierce feedback

---------

Co-authored-by: pierce <48131946+pierce314159@users.noreply.github.com>