Closes #2833: CI failures due to python `3.x/3.12` (#2834)
The python `3.x` job in the CI started pulling down `3.12`. This caused the CI to break. I imagine there's just some wheels or deps that aren't python `3.12` compatible yet. We saw something similar for python `3.11` a while back. This PR (closes #2833) fixes this by removing the `3.x` job for the time being and specfically checking `3.11`. We'll revisit once the dust settles a bit and try to add back the `3.12` and `3.x` checks

Co-authored-by: Pierce Hayes <pierce314159@users.noreply.github.com>