Closes #2843: Add python 3.12/3.x check to CI
This PR (closes #2843) adds back the 3.12/3.x python checks to the CI.

Unpinning tables and pyarrow seems to fix our dependency issues with the 3.12 check. These are the only two deps that have an upper bound so arkouda should be compatible with all the latest versions of our dependencies. I need to double check if the arrow version in the makefile has to match the one in yaml files