Closes #2838: Expand dataframe merge functions to accept multiple columns (#2848)
This PR (closes #2838) expands the dataframe merge functions to act on multiple columns. When no value is provided for `on`, it defaults to the intersection of the columns of the left and right dataframe. `inner_join_merge` and `right_join_merge` were turned into helper functions that aren't exposed to the user to more closely match the pandas merge functionality where these are only avialble through `merge`

Co-authored-by: Pierce Hayes <pierce314159@users.noreply.github.com>