Closes #2863: Implement `histogramdd` (#2868)
* Closes #2863: Implement `histogramdd`

This PR (closes #2862) adds the logic for a histogram with an arbitrary nubmer of dimensions. This mimics `np.histogramdd`. Added tests against the numpy version to ensure we have the same resulting histogram when acting on the same arrays.

This is based on the existing single dimensional histogram, so it has 3 different implementations based on number of bins to take advantage of the time / space tradeoff when total number of bins is relatively small. I tested all three versions against numpy with and without gasnet enabled

* add docstring

* changed list equals to np.allclose

---------

Co-authored-by: Pierce Hayes <pierce314159@users.noreply.github.com>