Closes #2873: Short strings optimization for multicol groupby (#2877)
* Closes #2873: Short strings optimization for multicol groupby

There's an optimization when calling groupby on a single array of small strings that skips the hash and operates on the byes directly. This PR (closes #2873) applies that optimization for multicolumn groupby where all the strings are small

NOTE: This only applies when the total number of bits being passed to radix sort is `<=128` (equivalently `totalDigits <= 8`). Otherwise all the columns are hashed and XORed together on a stride. So while this has less comm, it doesn't apply to as many cases as I'd like. Next steps are to try and improve the cause where `totalDigits>8`

* only cleanup strnames if we are actually replacing strings, this resolves unknown sym error on delete

---------

Co-authored-by: Pierce Hayes <pierce314159@users.noreply.github.com>