Fix string read bug for large Parquet files (#2938)
When the size of the number of rows in a parquet string column
exceeds the batch size, this can sometimes give incorrect results
since the readBatch function utilizes the definition level in a
unique way as opposed to other types, causing false positives for
whether or not a string was read. In order to resolve this, I am
reverting the batch size reading, which will reduce performance,
but since we are planning on reworking string reads anyway, this
isn't that big of a loss.