Fixes #3035 - Inconsistent results when broadcasting with empty segments (#3039)
* Fixes #3035 - Inconsitent results when broadcasting with empty segments

This PR (fixes #3035) a race condition bug that caused inconsistent results when broadcasting with empty segments (though both results given were incorrect).

We get around this by masking out the segments and values associated with the empty segments. We only do this when empty segments exist, so it shouldn't affect the performance of the common case too much but will give correct results in the empty seg case.

* had to add in empty check to avoid a segfault on one of the coargtests

---------

Co-authored-by: Tess Hayes <stress-tess@users.noreply.github.com>