Closes #3071: Add permutation to our generators (#3078)
* Closes #3071: Add permutation to rng

This PR (closes #3071) adds permutation to `rng`.

The array version of the chpl `permute` proc requires idxType must be coercible from this stream’s eltType, so instead we use the domain version of `permute` and use that to gather the permuted vals. But we're not missing out on much because the array version seems to do the same thing but without aggregation

* update in response to PR feedback and added float test

---------

Co-authored-by: Tess Hayes <stress-tess@users.noreply.github.com>