Part of #3088: Generate seed for sparse sum test (#3090)
This PR is more of a bandaid since I can't reproduce the error in #3088. I added randomly generated seeds that get printed to stdout when the test fails. This will make it easier to reproduce when it pops back up in the nightly

In the meantime I switched the sort workflow to be the default, so the merge workflow is opt in

Co-authored-by: Tess Hayes <stress-tess@users.noreply.github.com>