Fixes #3087: Failure reading non-float parquet columns with null values (#3094)
* Fixes #3087: Failure reading non-float parquet columns with null values

A user encountered this bug while reading the nyc taxi cab data. The pq file contained an int64 columns with null values. Pandas converted it into a float64 column and replaced the null values with `NaN`.

Properly handling null values is slower than the regular workflow since we can't read in batches. Since non-float columns having nulls is a fairly rare case, I decided to default to the faster way so the common case wouldn't take a performance hit. In the rare case, we fail with an informative error message that specifies which flag needs to be set to handle the null values. Rerunning with that flag behaves correctly and returns an array that matches pandas

* remove useless and confusing set equal to NAN lines

---------

Co-authored-by: Tess Hayes <stress-tess@users.noreply.github.com>