Fixes #3088: sparse sum nightly failures
This PR fixes #3088. In order to reproduce it I need to set the `ARKOUDA_DEVELOPER` flag to enable runtime checks, which I should've figured out sooner... Once we do this we get the following error message

```chapel
$CHPL_HOME/modules/internal/ChapelArray.chpl:1164: error: halt reached - array slice out of bounds
note: slice index was 804..815 but array bounds are 0..814
```

So I made all the loops serially and printed out the domains to find where the off by one was happening. This gives the right result with both `ARKOUDA_DEV` and `CHPL_DEV` set for the set seed and random. I tested this with various numbers of locales