Closes #3115: Fix non-deterministic sparse_sum failure
This PR (closes #3115) fixes the non-determinstic `sparse_sum_helper` failure. I was seeing runs where both the `if` and the `if else` would execute. Since the fetch conditions are mutually exclusive, we can use an `if` for both. I didn't realize that `on` statements were non-blocking, so there's pontential for them toexecute concurrently on different locales. Since my logic requires the previous locale to finish before the next begins, I've added an atomic helper to enforce this (thanks jeremiah for helping with that!)