Avoid a warning about mismatched parSafe settings for list initialization (#3149)
* Avoid a warning about mismatched parSafe settings for list initialization

The next release of Chapel will add a warning when initializing a list from
another list with a different value for `parSafe`, because it tripped up another
user.  This warning triggered for the creation of a new MessageArgs in
parseMessageArgs, because the MessageArgs type relies on the default parSafe
setting but the list being passed in was intentionally parSafe when it was
created (to allow for populating it more rapidly).

Rather than silence the warning for the program as a whole (and risk not
catching cases that did want the parSafe setting that is being ignored), add an
explicit initializer to MessageArgs that takes in a parallel safe list, copying
the contents rather than trying to use the list in the field's initialization.

This should not have a performance impact, as the list's contents are still only
being copied once.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>

* Drop `init this`, it's not compatible with 1.31 and it may be unnecessary

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>