Fixes #3158: `Dataframe` containing a `Segarray .__str__()` bug
This PR (closes #3158) fixes bug when printing a dataframe containing a segarray. To handle this I added an `nbytes` attribute to segarray which just wraps the underlying values. I also modified the series init to not assume that when `data` is a list that it contains 2 elements (the first being the data and the second being the index) when the `index` is provided.