Closes #3148: Update IO functions to always return a dictionary
Our IO functions used to return a dictionary if multiple columns or datasets were provided and if only was provided to just return the object. It has been requested by users that our IO functions always return a dictionary for consistency sake. This PR (closes #3148) makes the changes necessary for this and also updates all tests that relied on the old functionality