Fixes #3185: strings broadcast bug (#3210)
This PR fixes #3185 by removing the custom string permutation logic in favor of broadcasting the indices and then indexing into the segstring. I realized that we overcomplicating things and that optimizing this solution would essientailly be duplicating the segstring indexing code. I don't know what lead me to take this approach in the first place, but this is much cleaner and makes use of existing well-tested and optimized functions

Co-authored-by: Tess Hayes <stress-tess@users.noreply.github.com>