Fixes #3074: Bug reading segarrays from parquet files
We were seeing a bug where segarrays containing empty segments would fail if they exceeded ~700000 values. It seems our c++ code was mishandling the empty segments and this only appeared when values were large enough (my running theory is it has to do with the batch sizes). Reworking the logic to properly account for empty segment fixes #3074. Verified that when empty segs and nans are both present this is still properly handled