Reduce file size for csvIO benchmark (#3239)
This benchmark is timing out in multilocale runs due to bad performance
of the current implementation. There is a PR up
(https://github.com/Bears-R-Us/arkouda/pull/3238) which will improve
this implementation, but it may take a while to merge based on feedback.
Therefore we will reduce the file size here and bring it back up later
once the performance improves

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>