Improve the CSV Read performance with new impl (#3238)
* Improve the CSV Read performance with new impl

This improves the CSV read performance, especially in multilocale and
multi file settings where the performance previously was abysmal. See PR
for more info.

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>

* Don't use readDeserializable for 1.31 compat

Add it back once 1.31 support is dropped

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>

* Bring back up the problem size for the csv benchmark

This problem size had been reduced since it was timing out with the old
implementation. Now with this new one, everything should be faster and
should not time out

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>

* Not use try! in regex initializer

Also not compile regex when the type isn't string

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>

* Use readDeserializable interface

Now that we've dropped 1.31 support, we can add this back

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>

---------

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>