Part of #3229: Re-enable find implementation of `indexof1d`
This PR is part of #3229. Re-enable the `find` implementation. I modified it to have a more optimized way of finding all occurences (which will hopefully also solve the issue we were seeing in the CI). While I'm not sure if this is the best way of calculating this, it's no longer the bottleneck and it gets better performance than the current implementation.

Previously I added seeds to the `indexof1d` test, so it will be easier to reproduce any failures. I'm not closing the issue yet since there's still some code cleanup to be done, and I want to leave exisiting code to make it easier to fall back to a similar approach in the event that the `find` way still has issues.