Be aware of starting line index for files
Some files have data meant to be distributed across multiple locales
therefore we have to be aware of the lower bound of the line indices to
be read in. In order to do this, we skip over the lines until this lower
bound is reached. Since the domains are continuous, we don't need to
worry about upper bounds.

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>